refactor(domain): extract common methods to abstract class
