+++ src/main/java/baseball/domain/Answer.java
@@ -2,38 +2,15 @@ package baseball.domain;
 
 import static baseball.util.BaseballNumberGenerator.generateRandomNumber;
 import static baseball.util.Constants.BASEBALL_NUMBER_SIZE;
-import static baseball.util.ErrorMessage.NUMBER_SIZE;
-import static baseball.util.ErrorMessage.NUMBER_DISTINCT;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
-public class Answer {
-    private final List<BaseballNumber> numbers;
+public class Answer extends BaseballNumbers{
 
     public Answer(List<BaseballNumber> numbers) {
-        validate(numbers);
-        this.numbers = numbers;
-    }
-
-    private void validate(List<BaseballNumber> numbers) {
-        isValidSize(numbers);
-        isDistinct(numbers);
-    }
-
-    private void isValidSize(List<BaseballNumber> numbers) {
-        if (numbers.size() != BASEBALL_NUMBER_SIZE) {
-            throw new IllegalArgumentException(
-                    String.format(NUMBER_SIZE.message(), BASEBALL_NUMBER_SIZE));
-        }
-    }
-
-    private void isDistinct(List<BaseballNumber> numbers) {
-        if (numbers.stream().distinct().count() != numbers.size()) {
-            throw new IllegalArgumentException(NUMBER_DISTINCT.message());
-        }
+        super(numbers);
     }
 
     public static Answer generateAnswer() {
@@ -45,17 +22,5 @@ public class Answer {
             }
         }
         return new Answer(computer.stream().map(BaseballNumber::new).collect(Collectors.toList()));
-    }
-
-    public int indexOf(BaseballNumber number) {
-        return numbers.indexOf(number);
-    }
-
-    public boolean contains(BaseballNumber number) {
-        return numbers.contains(number);
-    }
-
-    public Stream<BaseballNumber> stream() {
-        return numbers.stream();
     }
 }
